# Conflicts:
#	Assets/InputSystem.meta
#	Assets/InputSystem/Controls.inputactions
#	Assets/InputSystem/Controls.inputactions.meta
#	Assets/InputSystem/InputSettings.asset
#	Assets/InputSystem/InputSettings.asset.meta
#	Assets/Materials/Skybox/Presets/PlayerSkybox_M.mat
#	Assets/Materials/VolumetricFog/FarFogPresets/PlayerFarVolumFog_M.mat
#	Assets/Materials/VolumetricFog/NearFogPresets/PlayerNearVolumFog_M.mat
#	Assets/Prefabs/Items/All Items Storage.prefab
#	Assets/Prefabs/Items/All Items Storage.prefab.meta
#	Assets/Prefabs/Items/SO/Clothing/Item_Gloves.asset
#	Assets/Prefabs/Items/SO/Clothing/Item_Hat.asset
#	Assets/Prefabs/Items/SO/Clothing/Item_Socks.asset
#	Assets/Prefabs/Items/Storage.prefab
#	Assets/Prefabs/Items/Storage.prefab.meta
#	Assets/Prefabs/Items/Storages/All Items Storage.prefab
#	Assets/Prefabs/Items/Storages/All Items Storage.prefab.meta
#	Assets/Scripts/InputSystem.meta
#	Assets/Scripts/InputSystem/Controls.inputactions
#	Assets/Scripts/InputSystem/Controls.inputactions.meta
#	Assets/Scripts/InputSystem/InputSettings.asset
#	Assets/Scripts/InputSystem/InputSettings.asset.meta
#	Assets/Scripts/World.cs
#	Assets/Settings/Input System.meta
#	Assets/Settings/Input System/Controls.inputactions
#	Assets/Settings/Input System/Controls.inputactions.meta
#	Assets/Settings/Input System/InputSettings.asset
#	Assets/Settings/Input System/InputSettings.asset.meta

+++ Assets/Scripts/Inventory/InventoryItem.cs
@@ -116,7 +116,7 @@ public abstract class InventoryItem : ScriptableObject
     [field: SerializeField] public DegradationType DegradeType { get; protected set; }
 
     [field: Tooltip("Скорость порчи предмета [ед/мин]"), Min(0.001f)]
-    [field: SerializeField] public float DegradationValue { get; private set; } = 1;
+    [field: SerializeField] public double DegradationValue { get; private set; } = 1;
 
     [field: Tooltip("Способы взаимодействия с предеметом"), Space(10)]
     [field: SerializeField] public ActionType Actions { get; protected set; }
